backend/requirements.render.txt
fastapi
uvicorn[standard]
python-dotenv
requests
azure-cognitiveservices-speech
openai
python-multipart
